Implement part of sign in function

+++ emart-back-end/user-service/src/main/java/com/fsd/userservice/dao/SignUpDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.NoRepositoryBean;
 
 @NoRepositoryBean
-public interface SignUpDao<T> extends JpaRepository<T, Long> {
+public interface SignUpDao<T> extends JpaRepository<T, Integer> {
 
 }
